New Counters
New multicoloured timers have been added. The counter was a PNG image which I edited via photoshop and changed the colour of. I then duplicated the code itself and just renamed all the necesarry functions to get it to work without causing any bugs or conflicts. The yellow one is for time and the purple is for jewels.  The red one hasn't been used yet but it could be used for displaying lives